requirements.txt
aiogram==3.0.0b7
openai==1.14.3
pymongo==4.7.2
python-dotenv==1.0.1
aiohttp==3.9.5